feat: Add guardian support to registration page
Adds a new Guardians table that creates a Many-To-Many relationship with the User table to link parents to children. It also links one-to-one with the User table to relate to a Parent User. This is intended to allow parents to transition into a more complete role if they like. This should make it easier to add parent specific features in the future. 

Currently parents are only shown on a child's profile. There are plans to add more views of this information that will be more helpful in the future.