chore: Cleanup guardian creation and namign
Re-work guardian creation so that the Guardian class drives creating the user and associating it to the Guardian object.

Now creating the guardian object and linked user are done separately from linking the Student user to the guardian so that the registration code is the same whether or not the guardian already existed.

This also renames some code.