fix: Ensure user is a guardian when attempting to edit
This fixes a server error caused by passing a user id for someone without guardian data. While we shouldn't ever be generating a link like this, it's possible to do so manually.

This also changes the redirect on missing (or incorrect) user to go to the list of guardians instead of the user page.