feat: Add event registration
Adds the ability for users to register for events. Registered users are shown on the event stats page visible to mentors. 

This mechanism is intended to support multiple time blocks for each event. Currently a single block for the whole duration is automatically created. 

Each block tracks whether the user is registered, and a comment. The comment field is intended for indicating if someone is arriving late, leaving early, etc.

Users names will show up as links on an event page if they have left a comment for the block. Clicking the link opens a modal with their comment.

Adds a new page visible to all users showing events that are open for registration and a link to the registration page for each event.